chore: cleanup and organizing
