+++ src/QueryBuilder/Documents/DesignDoc_QueryBuilder.md
