+++ src/QueryBuilder/components/QueryBuilderInput/queryBuilderInputTypes.ts
@@ -1,0 +1,5 @@
+import { SxProps } from '@mui/material';
+
+export interface BaseQueryBuilderInputFieldProps {
+  textFieldOverrides?: SxProps;
+}
